Implement deserialization via bytecode
