--- serde_arrow/src/arrow/test/mod.rs
@@ -1,1 +1,0 @@
-mod sinks;
