+++ serde_arrow/src/test/mod.rs
@@ -1,4 +1,2 @@
 mod api_chrono;
 mod error;
-
-pub mod utils;
