--- serde_arrow/src/test/utils.rs
@@ -1,11 +1,0 @@
-use crate::internal::{error::Result, event::Event, source::EventSource};
-
-pub fn collect_events<'event, S: EventSource<'event> + 'event>(
-    mut source: S,
-) -> Result<Vec<Event<'static>>> {
-    let mut res = Vec::new();
-    while let Some(ev) = source.next()? {
-        res.push(ev.to_static());
-    }
-    Ok(res)
-}
