+++ serde_arrow/src/test_impls/mod.rs
@@ -9,6 +9,3 @@ mod r#struct;
 mod tuple;
 mod r#union;
 mod utils;
-
-// TODO: remove
-mod playground;
